mono-luck-chores/src/example.py
import os
from dotenv import load_dotenv
load_dotenv()

# <variable name> = os.getenv(<environment variable name>)
DB_HOST = os.getenv('DB_HOST')
DB_NAME = os.getenv('DB_NAME')
DB_USER = os.getenv('DB_USER')
DB_PORT = os.getenv('DB_PORT')
DB_PASSWORD = os.getenv('DB_PASSWORD')
DB_URL = os.getenv('DB_URL')