mono-luck-chores/requirements.txt
docx2pdf==0.1.8
python-docx==0.8.11
python-dotenv
psycopg2
pytz
